extract `stringifyNonstringColumns` to `db-commons`
